👷‍♂️: only update apt if debian

+++ src/rye/install.sh
@@ -1,8 +1,10 @@
 #!/bin/sh
 set -e
 
-apt-get update
-apt-get install -y curl
+if [ -f /etc/debian_version ]; then
+  apt-get update
+  apt-get install -y curl
+fi
 
 if test -f /usr/local/share/rye/env; then
   echo "Rye already installed!"
